Create test_model_constraints.py and pull out two existing constraint tests
